requirements.txt
google-genai
python-dotenv
playwright
pydantic
jinja2
markdown
streamlit
pandas
Pillow